refactor(infrastructure): rename and update import paths for infrastructure module
